% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxfish.R
\name{boxfish}
\alias{boxfish}
\title{boxfish}
\usage{
boxfish()
}
\value{
a plot with grouped boxplots
}
\description{
This Shiny App will compare chemical amounts across categories
}
\examples{
\dontrun{mercfish()}
}
