% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{EPA Rivers and Streams Data}
\format{
A data frame 290 observations and 8 variables:
\describe{ **EDIT FIELDS BELOW FOR VARIABLES**
  \item{epaRegion}{The EPA Region in which the fish composite sample was collected}
  \item{state}{U.S. Postal Service 2-letter abbreviation for the state in which the fish composite sample was collected}
  \item{epaSampID}{Unique 6-digit number assigned by EPA to each fish composite sample}
  \item{amtMERC}{Concentration of the chemical}
  \item{amtPCB}{Concentration of the chemical}
  \item{amtPFA}{Concentration of the chemical}
  \item{species}{Generally accepted common name based on Nelson et al. (2004)}
  \item{feeder}{Classification as P=Predator or BD=Bottom-Dweller}
  ...
}
}
\source{
<https://riverstreamassessment.epa.gov/webreport/>
}
\usage{
data
}
\description{
From the EPA National Rivers and Streams Assessment 2018-2019
}
\keyword{datasets}
