% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterfish.R
\name{scatterfish}
\alias{scatterfish}
\title{scatterfish}
\usage{
scatterfish()
}
\value{
a scatterplot with colored groups
}
\description{
This Shiny App Will Perform SLR Between Chemical Concentrations and Color by Group
}
\examples{
\dontrun{scatterfish()}
}
